requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
pyrealsense2>=2.54.0
Pillow>=10.0.0
tqdm>=4.65.0
matplotlib>=3.7.0
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.1.0
